pyproject.toml
[project]
name = "python-module-tasks"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "xDepcio", email = "olek.drwal@gmail.com"},
]
dependencies = [
    "py-cpuinfo>=9.0.0",
    "psutil>=6.1.0",
    "requests>=2.32.3",
    "python-dotenv>=1.0.1",
    "pydantic>=2.9.2",
    "fastapi>=0.115.4",
    "typer>=0.12.5",
    "uvicorn>=0.32.0",
    "pytest>=8.3.3",
    "httpx>=0.27.2",
]
requires-python = ">=3.12"
readme = "README.md"
license = {text = "MIT"}

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"


[tool.pdm]
distribution = true

[tool.pdm.scripts]
task1_1 = "python3 -m python_module_tasks.task1.1.main"
task1_2 = "python3 -m python_module_tasks.task1.2.main"
task1_3 = "python3 -m python_module_tasks.task1.3.main"
task1_4 = "python3 -m python_module_tasks.task1.4.main"
task1_5 = "python3 -m python_module_tasks.task1.5.main"
task2 = "python3 -m python_module_tasks.task2.main"
start-api = "python3 -m python_module_tasks.advanced.api.main"
cli = "python3 -m python_module_tasks.advanced.client.cli"
